Move source-specific UDF rewrite functionality to datafusion-datasource
