+++ datafusion/datasource/src/projection.rs
@@ -17,6 +17,7 @@
 
 use std::sync::Arc;
 
+use crate::rewrite::rewrite_input_file_name_in_projection;
 use arrow::datatypes::{Schema, SchemaRef};
 use datafusion_common::{
     Result, ScalarValue,
@@ -26,7 +27,6 @@ use datafusion_physical_expr::{
     expressions::{Column, Literal},
     projection::{ProjectionExpr, ProjectionExprs},
 };
-use datafusion_physical_expr_adapter::schema_rewriter::rewrite_input_file_name_in_projection;
 use futures::{FutureExt, StreamExt};
 use itertools::Itertools;
